requirements.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
pdfkit==0.6.1
python-dotenv==0.14.0
python-http-client==3.3.1
sendgrid==6.4.7
starkbank-ecdsa==1.1.0
